Boilerplaterized character controller example.
